requirements.txt
pandas
beautifulsoup4
python-dotenv
selenium
pyyaml